perf: arm_pmuv3: Prepare for more than 32 counters
ANBZ: #25252
commit a4a6e2078d85a9d94bcc7eab77845cb8cd39f680 upstream.

Various PMUv3 registers which are a mask of counters are 64-bit
registers, but the accessor functions take a u32. This has been fine as
the upper 32-bits have been RES0 as there has been a maximum of 32
counters prior to Armv9.4/8.9. With Armv9.4/8.9, a 33rd counter is
added. Update the accessor functions to use a u64 instead.

Acked-by: Mark Rutland <mark.rutland@arm.com>
Tested-by: James Clark <james.clark@linaro.org>
Link: https://lore.kernel.org/r/20240731-arm-pmu-3-9-icntr-v3-2-280a8d7ff465@kernel.org

Update kvm_pmu_ops to use u64 for CONFIG_KVM_ARM_HOST_VHE_ONLY.
Link: https://gitee.com/anolis/cloud-kernel/pulls/6115

+++ arch/arm64/include/asm/kvm_host.h
@@ -1134,12 +1134,12 @@ void kvm_arch_vcpu_load_debug_state_flags(struct kvm_vcpu *vcpu);
 void kvm_arch_vcpu_put_debug_state_flags(struct kvm_vcpu *vcpu);
 
 #if IS_ENABLED(CONFIG_KVM)
-void kvm_set_pmu_events(u32 set, struct perf_event_attr *attr);
-void kvm_clr_pmu_events(u32 clr);
+void kvm_set_pmu_events(u64 set, struct perf_event_attr *attr);
+void kvm_clr_pmu_events(u64 clr);
 bool kvm_set_pmuserenr(u64 val);
 #else
-static inline void kvm_set_pmu_events(u32 set, struct perf_event_attr *attr) {}
-static inline void kvm_clr_pmu_events(u32 clr) {}
+static inline void kvm_set_pmu_events(u64 set, struct perf_event_attr *attr) {}
+static inline void kvm_clr_pmu_events(u64 clr) {}
 static inline bool kvm_set_pmuserenr(u64 val)
 {
 	return false;
@@ -1193,8 +1193,8 @@ bool kvm_arm_vcpu_stopped(struct kvm_vcpu *vcpu);
 
 #ifdef CONFIG_KVM_ARM_HOST_VHE_ONLY
 struct kvm_pmu_ops {
-	void (*set_pmu_events)(u32 set, struct perf_event_attr *attr);
-	void (*clr_pmu_events)(u32 clr);
+	void (*set_pmu_events)(u64 set, struct perf_event_attr *attr);
+	void (*clr_pmu_events)(u64 clr);
 	bool (*set_pmuserenr)(u64 val);
 	void (*vcpu_pmu_resync_el0)(void);
 };
